Projects/Project5_dashboard_project/requirements.txt
langchain
langchain-openai
langchain-community
langchain-core
langsmith
pandas
numpy
python-dotenv
openai
dash
plotly
